chore: update missing files
